+++ tests/test_interpolate.py
@@ -35,18 +35,19 @@ def test_laplace_interpolate():
     j = np.array([1, 0, 2, 1, 3, 2, 4])
     coo_content = (j, (i, j))
     data = np.array([1.0, np.nan, np.nan, np.nan, 5.0])
+    labels = np.zeros_like(data, dtype=int)
     with pytest.raises(ValueError, match="connectivity is not a square matrix"):
         con = sparse.coo_matrix(coo_content, shape=(4, 5)).tocsr()
-        interpolate.laplace_interpolate(data, con, use_weights=False)
+        interpolate.laplace_interpolate(data, con, labels, use_weights=False)
 
     expected = np.arange(1.0, 6.0)
     con = sparse.coo_matrix(coo_content, shape=(5, 5)).tocsr()
     actual = interpolate.laplace_interpolate(
-        data, con, use_weights=False, direct_solve=True
+        data, con, labels, use_weights=False, direct_solve=True
     )
     assert np.allclose(actual, expected)
 
     actual = interpolate.laplace_interpolate(
-        data, con, use_weights=False, direct_solve=False
+        data, con, labels, use_weights=False, direct_solve=False
     )
     assert np.allclose(actual, expected)